+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.7.6.dev
 * Extension: item verification on commandline
+* Improvement: catch lack of support of PERIOD in vobject <= 0.9.9
 
 ## 3.7.5
 * Add: [sharing] conversion_bday_summary_template (customize summary)